fix(windows): Fix MSVC compilation errors and warnings
Fix all Windows/MSVC build issues reported in GitHub Actions:

**Critical errors fixed:**
1. db.cpp: Wrap unistd.h with #ifndef _WIN32 (POSIX-only header)
2. influxdb.cpp: Add missing #include <chrono> for std::chrono
3. obj_save.cpp: Wrap chmod() calls with #ifndef _WIN32 (POSIX-only)

**Warnings fixed:**
1. action_targeting.h: Add static_cast<int> for size_t → int (C4267)
2. json_helpers.cpp: Fix shift warnings and size_t conversions (C4334, C4267)
   - Cast plane to unsigned int before 30-bit shift
   - Use unsigned int for flag_value intermediate storage
   - Replace ↔ with <-> to avoid UTF-8 in ASCII-only file
3. crud_handlers.cpp: Remove unused catch variable names (C4101)
   - Change catch (const json::parse_error& e) to catch (const json::parse_error&)

All changes are platform-safe:
- POSIX-specific code now wrapped with #ifndef _WIN32
- Explicit casts prevent data loss warnings
- Unused variables removed without affecting functionality

Verified: Linux build compiles successfully.

+++ src/engine/network/admin_api/json_helpers.cpp
@@ -30,8 +30,8 @@ json SerializeFlags(const FlagData& flagset)
 			if (plane_bits & (1U << bit))
 			{
 				// Encode as: (plane << 30) | (1 << bit)
-				int flag_value = (plane << 30) | (1U << bit);
-				flags_array.push_back(flag_value);
+				unsigned int flag_value = (static_cast<unsigned int>(plane) << 30) | (1U << bit);
+				flags_array.push_back(static_cast<int>(flag_value));
 			}
 		}
 	}
@@ -47,8 +47,8 @@ json SerializeBitvector(Bitvector bits, size_t plane)
 	{
 		if (bits & (1U << bit))
 		{
-			int flag_value = (plane << 30) | (1U << bit);
-			flags_array.push_back(flag_value);
+			unsigned int flag_value = (static_cast<unsigned int>(plane) << 30) | (1U << bit);
+			flags_array.push_back(static_cast<int>(flag_value));
 		}
 	}
 
@@ -56,7 +56,7 @@ json SerializeBitvector(Bitvector bits, size_t plane)
 }
 
 // ============================================================================
-// String Conversion (KOI8-R Б├■ UTF-8)
+// String Conversion (KOI8-R <-> UTF-8)
 // ============================================================================
 
 std::string Koi8rToUtf8(const char* koi8r)
